backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
email-validator
tensorflow==2.20.0
scikit-learn
joblib
numpy
pandas
psutil
bcrypt
python-jose[cryptography]
pyotp
qrcode
Pillow
requests
python-multipart